........
  r10706 | joost | 2008-04-18 22:14:21 +0200 (Fri, 18 Apr 2008) | 1 line
  
   * Removed "as" keyword from query, it is not supported by Firebird
........
  r10740 | joost | 2008-04-20 21:19:44 +0200 (Sun, 20 Apr 2008) | 1 line
  
   * Add sqQuoteFieldnames to postgresql connectionparameters
........
  r10743 | joost | 2008-04-20 22:34:04 +0200 (Sun, 20 Apr 2008) | 1 line
  
   * Handle String-Parameters as text parameters
........
  r10744 | joost | 2008-04-20 23:28:32 +0200 (Sun, 20 Apr 2008) | 1 line
  
   * Added support for stColumns to GetSchemaInfoSQL
........
  r10752 | joost | 2008-04-21 21:40:05 +0200 (Mon, 21 Apr 2008) | 1 line
  
   * Refactored test
........
  r10763 | joost | 2008-04-22 22:38:48 +0200 (Tue, 22 Apr 2008) | 1 line
  
   * Base indexes on AnsiCompareStr/AnsiCompareText and modified tests
........
  r10779 | joost | 2008-04-23 23:26:31 +0200 (Wed, 23 Apr 2008) | 1 line
  
   * Mysql servers before version 5.0.3 removes trailing spaces on varchar fields, adapted testvalues to take this into account
........
  r10786 | joost | 2008-04-24 22:33:44 +0200 (Thu, 24 Apr 2008) | 1 line
  
   * Refactored test-values, removed duplicate and re-enabled a string-testvalue containing quotes
........
  r10807 | joost | 2008-04-26 22:43:44 +0200 (Sat, 26 Apr 2008) | 3 lines
  
   * Fixed warning
   * Use consistent test data between 32-bit and 64-bit
   * Fix range-overflow problem in building indexes
........
  r10812 | joost | 2008-04-27 10:37:15 +0200 (Sun, 27 Apr 2008) | 1 line
  
   * Patch from Luk Vandelaer to use username, password and pagesize on creation of a database
........
  r10814 | joost | 2008-04-27 13:56:05 +0200 (Sun, 27 Apr 2008) | 1 line
  
   * Clean up TestScript-test properly
........
  r10817 | joost | 2008-04-27 17:46:55 +0200 (Sun, 27 Apr 2008) | 1 line
  
   * Map string-fields with a size above dsMaxStringSize to ftMemo-fields
........
  r10818 | joost | 2008-04-27 18:13:11 +0200 (Sun, 27 Apr 2008) | 1 line
  
   * Improved r10817
........
  r10819 | joost | 2008-04-27 18:14:04 +0200 (Sun, 27 Apr 2008) | 1 line
  
   * Allow a size (amount of decimals) of 0 for TBCDFields
........
  r10821 | joost | 2008-04-27 22:13:49 +0200 (Sun, 27 Apr 2008) | 1 line
  
   * Correct a test for mysql version 5 (server)
........
  r10822 | joost | 2008-04-27 22:15:46 +0200 (Sun, 27 Apr 2008) | 1 line
  
   * Fixes for ftCurrency fields
........
  r10824 | joost | 2008-04-27 22:24:29 +0200 (Sun, 27 Apr 2008) | 1 line
  
   * Reverted accidental commit in r10821
........
  r10839 | joost | 2008-04-29 23:54:45 +0200 (Tue, 29 Apr 2008) | 1 line
  
   * Commit the DDL-transaction in the TestScript test. (interbase only)
........
  r10844 | joost | 2008-04-30 19:32:12 +0200 (Wed, 30 Apr 2008) | 1 line
  
   * Avoid second exception while handling the first
........

git-svn-id: branches/fixes_2_2@10935 -

+++ packages/fcl-db/tests/sqldbtoolsunit.pas
@@ -94,6 +94,7 @@ implementation
 
 procedure TSQLDBConnector.CreateFConnection;
 var i : TSQLDBTypes;
+    t : integer;
 begin
   for i := low(DBTypesNames) to high(DBTypesNames) do
     if UpperCase(dbconnectorparams) = DBTypesNames[i] then sqldbtype := i;
@@ -102,6 +103,13 @@ begin
     
   if SQLDbType = MYSQL40 then Fconnection := tMySQL40Connection.Create(nil);
   if SQLDbType = MYSQL41 then Fconnection := tMySQL41Connection.Create(nil);
+  if SQLDbType in [mysql40,mysql41] then
+    begin
+    // Mysql versions prior to 5.0.3 removes the trailing spaces on varchar
+    // fields on insertion. So to test properly, we have to do the same
+    for t := 0 to testValuesCount-1 do
+      testStringValues[t] := TrimRight(testStringValues[t]);
+    end;
   if SQLDbType = MYSQL50 then Fconnection := tMySQL50Connection.Create(nil);
   if SQLDbType in MySQLdbTypes then
     FieldtypeDefinitions[ftLargeint] := 'BIGINT';
@@ -209,7 +217,7 @@ begin
           begin
           sql := sql + ',F' + Fieldtypenames[FType];
           if testValues[FType,CountID] <> '' then
-            sql1 := sql1 + ',''' + testValues[FType,CountID] + ''''
+            sql1 := sql1 + ',''' + StringReplace(testValues[FType,CountID],'''','''''',[rfReplaceAll]) + ''''
           else
             sql1 := sql1 + ',NULL';
           end;
@@ -227,26 +235,32 @@ end;
 
 procedure TSQLDBConnector.DropNDatasets;
 begin
-  try
-    if Ftransaction.Active then Ftransaction.Rollback;
-    Ftransaction.StartTransaction;
-    Fconnection.ExecuteDirect('DROP TABLE FPDEV');
-    Ftransaction.Commit;
-  Except
-    if Ftransaction.Active then Ftransaction.Rollback
-  end;
+  if assigned(FTransaction) then
+    begin
+    try
+      if Ftransaction.Active then Ftransaction.Rollback;
+      Ftransaction.StartTransaction;
+      Fconnection.ExecuteDirect('DROP TABLE FPDEV');
+      Ftransaction.Commit;
+    Except
+      if Ftransaction.Active then Ftransaction.Rollback
+    end;
+    end;
 end;
 
 procedure TSQLDBConnector.DropFieldDataset;
 begin
-  try
-    if Ftransaction.Active then Ftransaction.Rollback;
-    Ftransaction.StartTransaction;
-    Fconnection.ExecuteDirect('DROP TABLE FPDEV_FIELD');
-    Ftransaction.Commit;
-  Except
-    if Ftransaction.Active then Ftransaction.Rollback
-  end;
+  if assigned(FTransaction) then
+    begin
+    try
+      if Ftransaction.Active then Ftransaction.Rollback;
+      Ftransaction.StartTransaction;
+      Fconnection.ExecuteDirect('DROP TABLE FPDEV_FIELD');
+      Ftransaction.Commit;
+    Except
+      if Ftransaction.Active then Ftransaction.Rollback
+    end;
+    end;
 end;
 
 function TSQLDBConnector.InternalGetNDataset(n: integer): TDataset;
@@ -271,14 +285,17 @@ end;
 
 destructor TSQLDBConnector.Destroy;
 begin
-  try
-    if Ftransaction.Active then Ftransaction.Rollback;
-    Ftransaction.StartTransaction;
-    Fconnection.ExecuteDirect('DROP TABLE FPDEV2');
-    Ftransaction.Commit;
-  Except
-    if Ftransaction.Active then Ftransaction.Rollback
-  end;
+  if assigned(FTransaction) then
+    begin
+    try
+      if Ftransaction.Active then Ftransaction.Rollback;
+      Ftransaction.StartTransaction;
+      Fconnection.ExecuteDirect('DROP TABLE FPDEV2');
+      Ftransaction.Commit;
+    Except
+      if Ftransaction.Active then Ftransaction.Rollback
+    end; // try
+    end;
   inherited Destroy;
 
   FreeAndNil(FQuery);
